+++ 500g/Rata2i1/src/util.cpp
@@ -1,8 +1,8 @@
 //class with helper methods for other classes
 #include "util.h"
 
-int util::average(int values[], int size){
-    int sum = 0;
+double util::average(int values[], int size){
+    double sum = 0;
     for(int i = 0; i < size; i++){
         sum += values[i];
     }
